Added helper scripts for development within docker
Two helper scripts added for development within docker
containers:

- docker/build.sh: builds the et-plaform:latest docker
  image using sensible defaults.

- docker/run.sh: launches et-plaform:latest on a docker
  container, also with sensible defaults.

The run.sh script clones the user invoking the script
into the newly launched container. Then an optional
command is executed in the container using the
credentials of the invoking user and the current
directory.

Should the invoking user not provide any command then an
interactive shell is open.