Avoid docker rebuilds when developers install optional tools
Before this patch even small changes on comments within
docker/Dockerfile would make the docker builder to be unable to
cache the bulk of the build.

With this patch:

- The .dockerignore got the path for Dockerfile and other files
  updated. This was needed for the "COPY . ." to be cached when
  just small changes in the docker folder invalidated the cached
  results. It seems like when the docker files were moved to the
  docker folder the paths in .dockerignore were left outdated.
  This patch fixes those paths.

- The installed ubuntu packages are now split in two sets for
  handling differently stric build dependencies and optional
  packages that just make life easier for developers.

The optional packages are installed at the end of docker build, so
updates of the list of optional packages do not invalidate the
cached build results.

With these changes, when a new developer comes and adds emacs
as an editor to the list of optional packages he will be able
to get his wish granted without retriggering a full rebuild of
the docker for everyone.